make rust test pass
